fix(bank-notification): change @Table annotation to reactive and change ID Long -> Integer.

+++ services/bank-notification/src/main/java/am/banking/system/notification/model/repository/NotificationRepository.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
  * Time: 00:47:23
  */
 @Repository
-public interface NotificationRepository extends ReactiveCrudRepository<Notification, Long> {
+public interface NotificationRepository extends ReactiveCrudRepository<Notification, Integer> {
 }
